pyproject.toml
[project]
name = "plateau-gradio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "gradio>=6.12.0",
    "pillow>=12.2.0",
    "ultralytics>=8.4.39",
]
